requirements-deploy.txt
# Dependencias básicas de Flask
Flask==3.1.0
gunicorn==23.0.0
Werkzeug==3.1.3
Jinja2==3.1.5
MarkupSafe==3.0.2
itsdangerous==2.2.0
click==8.1.8
blinker==1.9.0

# Dependencias para detección de caras
opencv-python-headless==4.10.0.84
numpy>=1.26.0,<2.0.0
Pillow>=10.0.0,<11.0.0

# Dependencias adicionales
colorama==0.4.6
packaging==24.2
